+++ tests/test_plotpostprocess.py
@@ -7,8 +7,8 @@ from clease import Evaluate
 def test_plot_fit(bc_setting):
     predict = {
         "title": "plot_FIT_TEST",
-        "xlabel": "E_DFT (eV/atom)",
-        "ylabel": "DFT_FIT",
+        "xlabel": "E_CE (eV/atom)",
+        "ylabel": "E_DFT (eV/atom)",
     }
     plot_args = {"title": predict["title"], "ylabel": predict["ylabel"]}
     evaluator = Evaluate(bc_setting, fitting_scheme="l2", alpha=1e-6)